MarbleRun keyboard layout for Microsoft Windows.
The MarbleRun.klc keyboard definition file can be opened with the
keyboard layout creator of the Microsoft PowerToys tools. It will
generate an installer which adds MarbleRun as a US language keyboard
to Microsoft Windows.
For convenience there's also a pre-generated installer included, not
recommended for general use because running programs from untrusted
sources is bad practice, securitywise.